Update JAXMONDPMESH.py
Fixed a bug where code didnt work for 1 particle. This had to do with F being 0 in the MainLoop, giving a nan in interpolinv(F). These nans are now set to 0.

Also included a new functionality where particles leaving the boundary can be set to have 0 mass, by toggling the global Boolean CheckBoundary